Configuration Refactor
This commit refactors the entire sg-base role to use the newest
configuration refactor.  This refactor includes exciting new features
such as CARP, cleaner ways to specify multiple networks, and more!

The most notable change with this commit is that the gateway now
implements fairly advanced firewalling based on what services are
loaded.  Ideally in a future version this will be factored out to the
individual roles, but right now the template logic to do so would make
this commit too complex to review.

This refactor includes significant work from @m-wynn.